Refactor reflow_table into smaller helpers
Extract row parsing, width calculation and separator insertion into `reflow.rs`. This keeps `reflow_table` focused on orchestration and improves readability.
